+++ src/main/java/domain/WinningStats.java
@@ -7,14 +7,14 @@ import java.util.Map;
 import java.util.TreeMap;
 
 public class WinningStats {
-    private final Map<Ranking, Integer> prizeMap = new TreeMap<>();
-    private final int winningPrize;
+    private final Map<Ranking, Long> prizeMap = new TreeMap<>();
+    private final long winningPrize;
 
-    public WinningStats(int winningPrize) {
+    public WinningStats(long winningPrize) {
         this.winningPrize = winningPrize;
     }
 
-    public void addStats(Ranking ranking, int num) {
+    public void addStats(Ranking ranking, long num) {
         if (ranking != Ranking.NONE) {
             prizeMap.put(ranking, num);
         }
